+++ source/common/quic/quic_transport_socket_factory.cc
@@ -13,32 +13,6 @@
 namespace Envoy {
 namespace Quic {
 
-Network::DownstreamTransportSocketFactoryPtr
-QuicServerTransportSocketConfigFactory::createTransportSocketFactory(
-    const Protobuf::Message& config, Server::Configuration::TransportSocketFactoryContext& context,
-    const std::vector<std::string>& /*server_names*/) {
-  auto quic_transport = MessageUtil::downcastAndValidate<
-      const envoy::extensions::transport_sockets::quic::v3::QuicDownstreamTransport&>(
-      config, context.messageValidationVisitor());
-  auto server_config = std::make_unique<Extensions::TransportSockets::Tls::ServerContextConfigImpl>(
-      quic_transport.downstream_tls_context(), context);
-  // TODO(RyanTheOptimist): support TLS client authentication.
-  if (server_config->requireClientCertificate()) {
-    throwEnvoyExceptionOrPanic("TLS Client Authentication is not supported over QUIC");
-  }
-
-  auto factory = std::make_unique<QuicServerTransportSocketFactory>(
-      PROTOBUF_GET_WRAPPED_OR_DEFAULT(quic_transport, enable_early_data, true),
-      context.statsScope(), std::move(server_config));
-  factory->initialize();
-  return factory;
-}
-
-ProtobufTypes::MessagePtr QuicServerTransportSocketConfigFactory::createEmptyConfigProto() {
-  return std::make_unique<
-      envoy::extensions::transport_sockets::quic::v3::QuicDownstreamTransport>();
-}
-
 Network::UpstreamTransportSocketFactoryPtr
 QuicClientTransportSocketConfigFactory::createTransportSocketFactory(
     const Protobuf::Message& config,
@@ -84,9 +58,6 @@ std::shared_ptr<quic::QuicCryptoClientConfig> QuicClientTransportSocketFactory::
   // Return the latest crypto config.
   return crypto_config_;
 }
-
-REGISTER_FACTORY(QuicServerTransportSocketConfigFactory,
-                 Server::Configuration::DownstreamTransportSocketConfigFactory);
 
 REGISTER_FACTORY(QuicClientTransportSocketConfigFactory,
                  Server::Configuration::UpstreamTransportSocketConfigFactory);
